feat: Add booking Excel import/export functionality and update user ID type to Integer.

+++ src/main/java/com/example/booking_tour/dto/UserUpdateRequest.java
@@ -7,25 +7,25 @@ import lombok.Data;
 
 @Data
 public class UserUpdateRequest {
-    
-    private Long id;
-    
+
+    private Integer id;
+
     @NotBlank(message = "Name is required")
     @Size(min = 2, max = 100, message = "Name must be between 2 and 100 characters")
     private String name;
-    
+
     @NotBlank(message = "Email is required")
     @Email(message = "Email should be valid")
     private String email;
-    
+
     @NotBlank(message = "Phone is required")
     @Size(min = 10, max = 15, message = "Phone must be between 10 and 15 characters")
     private String phone;
-    
+
     private String avatar;
-    
+
     @NotBlank(message = "Role is required")
     private String role;
-    
+
     private String provider;
 }
